lambda/requirements.txt
aws-lambda-powertools==3.19.0
beautifulsoup4==4.13.4
Mastodon.py==2.1.2
openai==1.100.2
requests==2.32.5